move drawing functions from shared/ to smokeview/
There are some drawing functions that have been moved into shared/.
These functions are moved into smokeview/ as the are only used in
the smokeview program itself. This removes the GLUT include as well.

Currently there is still a dependency on OpenGL in shared/ due to some
textures in readcad.c, but no dependency on GLUT.
